core: reject type-incorrect state handling
Store a UDP source port in the parsed entropy value rather than the entropy-header pointer. Preserve the pointer type for deferred atomic state and declare the read-write lock initializer in its header.

Add a deterministic strict-core build that recompiles both engine variants with implicit declarations and integer conversions treated as errors, preventing these type violations from returning unnoticed.

+++ uet_api.c
@@ -2719,7 +2719,7 @@ static uet_ses_rc_t uet_rx_atomic_req_pkt(
 	if (sync_defer_atomic) {
 		entry->atomic_parms.opcode = UET_AMO_SUM;
 		entry->atomic_parms.data_type = UET_TYPE_UINT64;
-		entry->atomic_parms.addr = (uint64_t) addr;
+		entry->atomic_parms.addr = addr;
 		entry->atomic_parms.data = data;
 	} else
 		__atomic_fetch_add(addr, data, __ATOMIC_SEQ_CST);
